backend/requirements.txt
Django==6.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
gunicorn==25.3.0
psycopg>=3.0,<4.0